+++ packages/qwik-city/middleware/request-handler/request-handler.ts
@@ -1,8 +1,7 @@
 import { loadRoute } from '../../runtime/src/library/routing';
-import { loadEndpointResponse } from './endpoint-handler';
+import { loadUserResponse } from './user-response';
 import type { QwikCityRequestContext } from './types';
 import { ROUTE_TYPE_ENDPOINT } from '../../runtime/src/library/constants';
-import { checkPageRedirect } from './redirect-handler';
 import type { RenderToStringResult, StreamWriter } from '@builder.io/qwik/server';
 import { getQwikCityUserContext } from './utils';
 
@@ -19,40 +18,34 @@ export async function requestHandler<T = any>(requestCtx: QwikCityRequestContext
       const { mods, params, route } = loadedRoute;
       const isEndpointOnly = route[3] === ROUTE_TYPE_ENDPOINT;
 
-      if (!isEndpointOnly) {
-        // content page, so check if the trailing slash should be redirected
-        const redirectResponse = checkPageRedirect(url, trailingSlash, response);
-        if (redirectResponse) {
-          // page redirect will add or remove the trailing slash depending on the option
-          return redirectResponse;
-        }
-      }
-
       // build endpoint response from each module in the hierarchy
-      const userResponseContext = await loadEndpointResponse(
+      const userResponse = await loadUserResponse(
         request,
         url,
         params,
         mods,
+        trailingSlash,
         isEndpointOnly
       );
 
-      if (userResponseContext.handler === 'page') {
-        return response(userResponseContext.status, userResponseContext.headers, async (stream) => {
-          const result = await render({
-            stream,
-            url: url.href,
-            userContext: getQwikCityUserContext(userResponseContext),
-          });
-          if (result && (typeof result as any as RenderToStringResult).html === 'string') {
-            stream.write((result as any as RenderToStringResult).html);
+      if (userResponse.type === 'endpoint') {
+        return response(userResponse.status, userResponse.headers, async (stream) => {
+          if (typeof userResponse.body === 'string') {
+            stream.write(userResponse.body);
           }
         });
       }
 
-      if (userResponseContext.handler === 'endpoint') {
-        return response(userResponseContext.status, userResponseContext.headers, async (stream) => {
-          stream.write(userResponseContext.body);
+      if (userResponse.type === 'page') {
+        return response(userResponse.status, userResponse.headers, async (stream) => {
+          const result = await render({
+            stream,
+            url: url.href,
+            userContext: getQwikCityUserContext(userResponse),
+          });
+          if (result && (typeof result as any as RenderToStringResult).html === 'string') {
+            stream.write((result as any as RenderToStringResult).html);
+          }
         });
       }
     }
